requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
pytest>=7.4.0
pytest-mock>=3.12.0
# For compatibility with LangChain and OpenAI
requests==2.31.0
# For Windows compatibility
colorama>=0.4.6
pdfkit==1.0.0
PyPDF2==3.0.1
